@@ -1,32 +1,22 @@
-[tool.black]
-line-length = 80
-target-version = ['py39']
+[build-system]
+requires = ["poetry-core"]
+build-backend = "poetry.core.masonry.api"
 
-[tool.codespell]
-# Ignore false-positive on 'Nam', found in tests
-ignore-words-list = "nam"
-
-[tool.isort]
-profile = "black"
-line_length = 80
-
-[tool.mypy]
-exclude = "/locales/"
-packages = "pycountry"
-
-[tool.poetry]
+[project]
 name = "pycountry"
 version = "24.6.2.dev0"
 description = "ISO country, subdivision, language, currency and script definitions and their translations"
-authors = ["Christian Theune <ct@flyingcircus.io>"]
+authors = [
+	{name="Christian Theune", email="ct@flyingcircus.io"},
+]
 maintainers = [
-	"Nate Schimmoller <nschimmo@gmail.com>",
-	"Zachary Ware <zachary.ware@gmail.com>",
-	]
+	{name="Nate Schimmoller", email="nschimmo@gmail.com"},
+	{name="Zachary Ware", email="zachary.ware@gmail.com"},
+]
 license = "LGPL-2.1-only"
 readme = "README.rst"
-repository = "https://github.com/flyingcircusio/pycountry"
 keywords = [ "country", "subdivision", "language", "currency", "iso", "3166", "639", "4217", "15924", "3166-1","3166-2", "3166-3"]
+requires-python = ">=3.10"
 classifiers = [
         "Development Status :: 5 - Production/Stable",
         "Intended Audience :: Developers",
@@ -35,22 +25,21 @@ classifiers = [
         "Operating System :: OS Independent",
         "Programming Language :: Python",
         "Programming Language :: Python :: 3",
-        "Programming Language :: Python :: 3.9",
         "Programming Language :: Python :: 3.10",
         "Programming Language :: Python :: 3.11",
         "Programming Language :: Python :: 3.12",
         "Programming Language :: Python :: 3.13",
+        "Programming Language :: Python :: 3.14",
         "Topic :: Software Development :: Internationalization",
         "Topic :: Software Development :: Localization"
         ]
 
-[tool.poetry.dependencies]
-python = ">=3.9"
+[project.urls]
+source = "https://github.com/pycountry/pycountry"
+issues = "https://github.com/pycountry/pycountry/issues"
+changelog = "https://github.com/pycountry/pycountry/blob/main/HISTORY.txt"
 
 [tool.poetry.group.dev.dependencies]
-# Note: black and isort are not listed here, as they don't need to be installed
-# separately.  pre-commit will take care of running them, and tracking the
-# version in use.
 mypy = ">1.7.1"
 pre-commit = ">3.4.0"
 pytest = ">7.1.3"
@@ -62,9 +51,27 @@ docutils = "<0.21"
 
 [tool.pytest.ini_options]
 pythonpath = "src"
-addopts = "--cov --cov-report=html --cov-report=term-missing"
+addopts = "--doctest-glob='*.rst' --cov --cov-report=html --cov-report=term-missing"
+doctest_optionflags = "NORMALIZE_WHITESPACE IGNORE_EXCEPTION_DETAIL"
 required_plugins = "pytest-cov>=4.0.0"
 
-[build-system]
-requires = ["poetry-core>=1.0.0"]
-build-backend = "poetry.core.masonry.api"
+[tool.ruff]
+# See https://docs.astral.sh/ruff/settings/ for documentation and other options
+line-length = 80
+
+[tool.ruff.lint]
+select = [
+  "E4",
+  "E7",
+  "E9",
+  "F",
+  "I",
+]
+
+[tool.codespell]
+# Ignore false-positive on 'Nam', found in tests
+ignore-words-list = "nam"
+
+[tool.mypy]
+exclude = "/locales/"
+packages = "pycountry"
