@@ -10,7 +10,7 @@ import os.path
 import shutil
 import subprocess
 
-REVISION = "v4.18.0"
+REVISION = "v4.20.1"
 
 data_dir = "parts/data"
 base_dir = os.path.join("src", "pycountry")
